rename fetchstore
